rename: 입력 모델을 한 패키지로 이동
